Finished editor.js generator
