+++ core-web/libs/sdk/uve/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/core/core.utils';
 export * from './lib/dom/dom.utils';
+export * from './lib/editor/public';
